add facility policy, add facility tests

+++ app/controllers/api/v1/facilities_controller.rb
@@ -1,16 +1,23 @@
 class Api::V1::FacilitiesController < ApplicationController
+  skip_before_action :authenticate_request, only: %i[index]
+  before_action :authorize_policy
   before_action :set_accommodation
   before_action :set_facility, only: %i[update destroy]
-  skip_before_action :authenticate_request, only: %i[index]
 
   def index
     @facilities = @accommodation.facilities.all
+
+    authorize @facilities
+
     render json: @facilities, status: :ok
   end
 
   # POST /api/v1/facilities
   def create
     @facility = @accommodation.facilities.new(facility_params)
+    
+    authorize @facility
+    
     if @facility.save
       render json: { data: @facility }, status: :created
     else
@@ -20,6 +27,8 @@ class Api::V1::FacilitiesController < ApplicationController
 
   # PATCH/PUT /api/v1/facilities/1
   def update
+    authorize @facility
+
     if @facility.update(facility_params)
       render json: { status: 'Update', data: @facility }, status: :ok
     else
@@ -29,6 +38,8 @@ class Api::V1::FacilitiesController < ApplicationController
 
   # DELETE /api/v1/facilities/1
   def destroy
+    authorize @facility
+
     if @facility.destroy!
       render json: { status: 'Delete' }, status: :ok
     else
@@ -51,5 +62,8 @@ class Api::V1::FacilitiesController < ApplicationController
     params.permit(:checkin_start, :checkin_end, :checkout_start, :checkout_end, :credit_card, :free_parking,
                   :wi_fi, :breakfast, :pets)
   end
-end
 
+  def authorize_policy
+    authorize Facility
+  end
+end
